pyproject.toml
[project]
name = "discord-bot"
version = "0.1.0"
dependencies = [
    "apscheduler==3.11.0",
    "discord-py>=2.5.2",
    "python-dotenv>=1.1.0",
    "pytz==2024.2",
    "scheduler>=0.8.8",
    "tlgbotfwk>=0.4.61",
]

[tool.pdm.dependencies]
discord.py = ">=2.3.2"
apscheduler = "*"
pytz = "*"
